pyproject.toml
[project]
name = "nonebot-adapter-feishu"
version = "2.6.2"
description = "feishu(larksuite) adapter for nonebot2"
authors = [{ name = "StarHeartHunt", email = "starheart233@gmail.com" }]
requires-python = ">=3.9, <4.0"
readme = "README.md"
license = { text = "MIT" }
keywords = ["bot", "feishu", "larksuite"]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Framework :: Robot Framework",
  "Framework :: Robot Framework :: Library",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
]
dependencies = ["nonebot2 ~=2.2", "cashews ~=7.0", "pycryptodome ~=3.18"]

[project.urls]
homepage = "https://feishu.adapters.nonebot.dev/"
repository = "https://github.com/nonebot/adapter-feishu"
documentation = "https://feishu.adapters.nonebot.dev/"

[dependency-groups]
dev = [
  "ruff>=0.11.0,<0.12",
  "nonemoji>=0.1.2,<0.2",
  "pre-commit>=4.0.0,<5",
  "nonebot2[fastapi, httpx]",
  "nonebug",
]
test = [
  "trio >=0.27.0",
  "nonebug >=0.4.1, <0.5.0",
  "wsproto >=1.2.0, <2.0.0",
  "werkzeug >=2.3.6, <4.0.0",
  "pytest-cov >=6.0.0, <7.0.0",
  "pytest-xdist >=3.0.2, <4.0.0",
  "coverage-conditional-plugin >=0.9.0, <0.10.0",
]
docs = ["nb-autodoc >=1.0.0a5, <2.0.0"]

[tool.uv]
default-groups = ["dev", "test", "docs"]

[tool.uv.sources]
nonebot2 = { git = "https://github.com/nonebot/nonebot2.git", branch = "master" }
nonebug = { git = "https://github.com/nonebot/nonebug.git", branch = "master" }

[tool.uv.build-backend]
module-name = "nonebot.adapters.feishu"
module-root = ""

[tool.pytest.ini_options]
addopts = ["--cov=nonebot/adapters/feishu", "--cov-report=term-missing"]
filterwarnings = ["error", "ignore::DeprecationWarning"]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.format]
line-ending = "lf"

[tool.ruff.lint]
select = [
  "F",     # Pyflakes
  "W",     # pycodestyle warnings
  "E",     # pycodestyle errors
  "I",     # isort
  "UP",    # pyupgrade
  "ASYNC", # flake8-async
  "C4",    # flake8-comprehensions
  "T10",   # flake8-debugger
  "T20",   # flake8-print
  "PYI",   # flake8-pyi
  "PT",    # flake8-pytest-style
  "Q",     # flake8-quotes
  "TID",   # flake8-tidy-imports
  "RUF",   # Ruff-specific rules
]
ignore = [
  "E402",   # module-import-not-at-top-of-file
  "UP037",  # quoted-annotation
  "RUF001", # ambiguous-unicode-character-string
  "RUF002", # ambiguous-unicode-character-docstring
  "RUF003", # ambiguous-unicode-character-comment
]

[tool.ruff.lint.isort]
force-sort-within-sections = true
known-first-party = ["nonebot", "tests/*"]
extra-standard-library = ["typing_extensions"]

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true

[tool.pyright]
pythonVersion = "3.9"
pythonPlatform = "All"
defineConstant = { PYDANTIC_V2 = true }
executionEnvironments = [
  { root = "./tests", extraPaths = [
    "./",
  ] },
  { root = "./" },
]

typeCheckingMode = "standard"
reportShadowedImports = false
disableBytesTypePromotions = true

[build-system]
requires = ["uv_build >=0.8.3, <0.9.0"]
build-backend = "uv_build"
